agent-runtime/requirements.txt
httpx==0.27.0
fastapi==0.109.2
uvicorn==0.27.1
apscheduler==3.10.4
anthropic==0.49.0
openai==1.66.3
google-genai==1.10.0
discord.py==2.3.2
structlog==24.1.0
